@@ -48,29 +48,29 @@
     "react-gcolor-picker": "^1.2.4",
     "react-hotkeys-hook": "^3.4.4",
     "react-moment": "^1.1.1",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
     "reactour": "^1.18.7",
-    "serve": "^12.0.1",
+    "serve": "^13.0.2",
     "sockette": "^2.0.6",
     "styled-components": "^5.3.3",
-    "typescript": "^4.4.4",
+    "typescript": "^4.6.2",
     "use-api-polling": "^0.1.4",
     "use-long-press": "^1.1.1",
     "web-vitals": "^2.1.4",
-    "workbox-background-sync": "^5.1.4",
-    "workbox-broadcast-update": "^5.1.4",
-    "workbox-cacheable-response": "^5.1.4",
-    "workbox-core": "^5.1.4",
-    "workbox-expiration": "^5.1.4",
-    "workbox-google-analytics": "^5.1.4",
-    "workbox-navigation-preload": "^5.1.4",
-    "workbox-precaching": "^5.1.4",
-    "workbox-range-requests": "^5.1.4",
-    "workbox-routing": "^5.1.4",
-    "workbox-strategies": "^5.1.4",
-    "workbox-streams": "^5.1.4",
-    "zustand": "^3.6.1"
+    "workbox-background-sync": "^6.5.1",
+    "workbox-broadcast-update": "^6.5.1",
+    "workbox-cacheable-response": "^6.5.1",
+    "workbox-core": "^6.5.1",
+    "workbox-expiration": "^6.5.1",
+    "workbox-google-analytics": "^6.5.1",
+    "workbox-navigation-preload": "^6.5.1",
+    "workbox-precaching": "^6.5.1",
+    "workbox-range-requests": "^6.5.1",
+    "workbox-routing": "^6.5.1",
+    "workbox-strategies": "^6.5.1",
+    "workbox-streams": "^6.5.1",
+    "zustand": "^3.7.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -123,42 +123,42 @@
   },
   "devDependencies": {
     "@babel/core": "^7.17.7",
-    "@storybook/addon-actions": "^6.4.9",
-    "@storybook/addon-essentials": "^6.4.9",
-    "@storybook/addon-links": "^6.4.9",
-    "@storybook/addons": "^6.4.9",
-    "@storybook/builder-webpack5": "^6.4.9",
-    "@storybook/manager-webpack5": "^6.4.9",
-    "@storybook/node-logger": "^6.4.9",
-    "@storybook/preset-create-react-app": "^4.0.0",
-    "@storybook/theming": "^6.4.9",
-    "@types/react-router-dom": "^5.3.2",
-    "@typescript-eslint/eslint-plugin": "^5.2.0",
-    "@typescript-eslint/parser": "^5.2.0",
+    "@storybook/addon-actions": "^6.4.19",
+    "@storybook/addon-essentials": "^6.4.19",
+    "@storybook/addon-links": "^6.4.19",
+    "@storybook/addons": "^6.4.19",
+    "@storybook/builder-webpack5": "^6.4.19",
+    "@storybook/manager-webpack5": "^6.4.19",
+    "@storybook/node-logger": "^6.4.19",
+    "@storybook/preset-create-react-app": "^4.0.2",
+    "@storybook/theming": "^6.4.19",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^5.15.0",
+    "@typescript-eslint/parser": "^5.15.0",
     "babel-eslint": "^10.1.0",
-    "concurrently": "^6.3.0",
+    "concurrently": "^7.0.0",
     "cross-env": "^7.0.3",
-    "electron": "^14.0.1",
-    "electron-builder": "^22.13.1",
+    "electron": "^17.1.2",
+    "electron-builder": "^22.14.13",
     "electron-devtools-installer": "git+https://github.com/MarshallOfSound/electron-devtools-installer.git",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^16.1.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-airbnb-typescript": "^16.1.2",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-config-react-app": "^7.0.0",
     "eslint-import-resolver-typescript": "^2.4.0",
     "eslint-plugin-flowtype": "^8.0.2",
     "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jest": "^25.2.2",
+    "eslint-plugin-jest": "^26.1.1",
     "eslint-plugin-json": "^3.0.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
+    "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.0.8",
     "gh-pages": "^3.2.0",
-    "prettier": "^2.5.1",
-    "react-git-info": "^2.0.0",
+    "prettier": "^2.6.0",
+    "react-git-info": "^2.0.1",
     "react-is": "^17.0.2",
-    "wait-on": "^6.0.0",
+    "wait-on": "^6.0.1",
     "webpack": "^5.70.0"
   },
   "resolutions": {
